ctx: refactor to use uint32_t instead of uint64_t for string hashes
For our parser needs squoze-6 is collision free on strings we care
about, this also reduces internal memory usage for some data-structures.

+++ src/color.c
@@ -848,7 +848,7 @@ mrg_color_parse_hex (CtxState *ctxstate, CtxColor *color, const char *color_stri
 int ctx_color_set_from_string (Ctx *ctx, CtxColor *color, const char *string)
 {
   int i;
-  uint64_t hash = ctx_strhash (string);
+  uint32_t hash = ctx_strhash (string);
 //  ctx_color_set_rgba (&(ctx->state), color, 0.4,0.1,0.9,1.0);
 //  return 0;
     //rgba[0], rgba[1], rgba[2], rgba[3]);
